fix: 修改消息处理

+++ iface/imessage.go
@@ -5,8 +5,8 @@ type IMessage interface {
 	SetMsgId(id uint16)
 	GetProtoId() uint16
 	SetProtoId(id uint16)
-	GetMsgLen() uint16
-	SetMsgLen(len uint16)
+	GetMsgLen() uint32
+	SetMsgLen(len uint32)
 	GetData() []byte
 	SetData(data []byte)
 }
